requirements.txt
pyrtlsdr
numpy
pyvjoy
setuptools